+++ osdep/macOS_menubar.m
@@ -18,9 +18,9 @@
 #include "config.h"
 #include "common/common.h"
 
-#import "macosx_menubar_objc.h"
-#import "osdep/macosx_application_objc.h"
-#include "osdep/macosx_compat.h"
+#import "macOS_menubar_objc.h"
+#import "osdep/macOS_application_objc.h"
+#include "osdep/macOS_compat.h"
 
 @implementation MenuBar
 {
