+++ osdep/path.h
@@ -18,7 +18,7 @@ typedef const char *(*mp_get_platform_path_cb)(void *talloc_ctx, const char *typ
 // Conforming to mp_get_platform_path_cb.
 const char *mp_get_platform_path_uwp(void *talloc_ctx, const char *type);
 const char *mp_get_platform_path_win(void *talloc_ctx, const char *type);
-const char *mp_get_platform_path_osx(void *talloc_ctx, const char *type);
+const char *mp_get_platform_path_macos(void *talloc_ctx, const char *type);
 const char *mp_get_platform_path_unix(void *talloc_ctx, const char *type);
 
 #endif
