+++ player/main.c
@@ -67,7 +67,7 @@ static const char def_config[] =
 ;
 
 #if HAVE_COCOA
-#include "osdep/macosx_events.h"
+#include "osdep/macOS_events.h"
 #endif
 
 #ifndef FULLCONFIG
