+++ wscript_build.py
@@ -522,10 +522,10 @@ def build(ctx):
         ( "osdep/android/strnlen.c",             "android"),
         ( "osdep/ar/HIDRemote.m",                "apple-remote" ),
         ( "osdep/glob-win.c",                    "glob-win32" ),
-        ( "osdep/macosx_application.m",          "cocoa" ),
-        ( "osdep/macosx_events.m",               "cocoa" ),
-        ( "osdep/macosx_menubar.m",              "cocoa" ),
-        ( "osdep/macosx_touchbar.m",             "macos-touchbar" ),
+        ( "osdep/macOS_application.m",           "cocoa" ),
+        ( "osdep/macOS_events.m",                "cocoa" ),
+        ( "osdep/macOS_menubar.m",               "cocoa" ),
+        ( "osdep/macOS_touchbar.m",              "macos-touchbar" ),
         ( "osdep/mpv.rc",                        "win32-executable" ),
         ( "osdep/path-macosx.m",                 "cocoa" ),
         ( "osdep/path-unix.c"),
